refactor: 💡 Rename the file, containing algorithm
Refactor the code to a more readable one. Document the code using
`jsDoc`. Rename the file containing the algorithm solution to
`two-colored-graph.js`, as it is platform agnostic. Since the desired
solution should have GUI, having a file named `main.js` might give false
impression that it is a starting point of the programme. In fact, we
will need a wrapper module arround it, to handle user input from the
GUI.